runtime/lambda-authorizer/requirements.txt
pyjwt==1.7.1
